Update README.md
replaced old links with newer ones. 
indicated when scripts are not used in current version manuscript + when I removed parts of code because it was no longer used.
wrote description for script that creates the data for the 80% training and 20% test set + indicate how this is handled in the creation of the descriptive statistics
indicates how to handle multiple ancestries in various scripts.